backend/requirements-development.txt
-r requirements.txt

watchdog[watchmedo]==2.1.5
pytest==7.3.1
pytest-django==4.5.2
factory-boy==3.2.1
pytest-factoryboy==2.5.1
Faker==18.6.0
